+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 13.34.0
+## RevenueCat SDK
+### 🐞 Bugfixes
+* Fix purchases-js Store mapping (#1155) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* [AUTOMATIC] Android 8.19.0 => 8.19.1 JS 1.3.0 => 1.5.0 (#1153) via RevenueCat Git Bot (@RCGitBot)
+
+## 13.33.0
+## RevenueCat SDK
+### 🐞 Bugfixes
+* Store `appUserId` in `localStorage` for web hybrid support (#1147) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* [AUTOMATIC] iOS 5.24.0 => 5.26.0 Android 8.17.1 => 8.19.0 (#1151) via RevenueCat Git Bot (@RCGitBot)
+
+## 13.32.0
+## RevenueCat SDK
+### 📦 Dependency Updates
+* [AUTOMATIC] iOS 5.23.0 => 5.24.0 JS 1.2.1 => 1.3.0 (#1144) via RevenueCat Git Bot (@RCGitBot)
+
 ## 13.31.0
 ## RevenueCat SDK
 ### 📦 Dependency Updates